Use size_t indices in CFX_WideTextBuf::Delete()
- Rename CFX_BinaryBuf::Delete() to avoid confusion.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/90691
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fxcrt/cfx_widetextbuf.h
@@ -24,7 +24,7 @@ class CFX_WideTextBuf final : public CFX_BinaryBuf {
   WideString MakeString() const;
 
   void AppendChar(wchar_t wch);
-  void Delete(int start_index, int count);
+  void Delete(size_t start_index, size_t count);
 
   CFX_WideTextBuf& operator<<(ByteStringView ascii);
   CFX_WideTextBuf& operator<<(const wchar_t* lpsz);
